examples/wordx/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[dotenv]>=2.5.0
python-dotenv>=1.0.0
agentx